fix landscape layout
In the landscape mode put the board on the right-hand side. As a
result it is now possible to, say, read a book in a browser, hit the
"Recent Apps" button and see the whole board on the thumbnail itself,
while previously the left-hand side of the board on the thumbnail was
obscured by the SCID logo.