remove Cursor support from DataBaseView
Completely remove internal use of content provider (it is still
available for external requests).

Since DataBaseView does not need to support Cursor interface we use
direct methods in DataBaseView, for example,
 String getPGN()
 boolean isFavorite()
 [...]
instead of
 Bundle respond(Bundle extras)

Remove unneeded caching of data (fixes several bugs caused by the use
of stale copies).

Simplify reload detection in GameListActivity.

If a filter is applied, the title displays game position and the whole
size of the current database view as well as position and the total
number of games in the file.

Toasts were added: Cannot go to the previous (next) game.