better handling of the search outcome
If there are many results, then
 show them as a list (preserve the current position if possible).

If there is only one result, then
 if it can be shown (the found id is in the current filter), then
  show it and preserve the current filter
 else
  reset filter and show the game.

If there is no results, then preserve the current filter and game.

In all the cases, then the currently shown game is among the found,
make sure that we either directly go to its untouched state (possibly
updating the title due to the changed filter), or we can return to
original game by canceling the game list.