add auto play
"Auto play interval" preference is added: if it is non-zero, then a
long click on next move button starts autoplay with this interval in
seconds.

One may argue that a separate button to start auto-play would be
better, but since users do not want smaller buttons, and going to the
last move in variation is easily done by clicking the move list, let
us try this approach (as a bonus, the preference makes the feature
easy to discover).