add option to report thinking time in study mode
The time is reported then user does a move on the board.

The human thinking timer is not reset on a click on next or previous
move button (only on doing the move on the board), thus it is possible
to review the previous position (say, to find what was just captured)
and do not affect the timer. If the move is wrong, the timer is also
not reset.