show progress in search during reading of the index file
Before this, in a huge database the first search is hung with 0
progress all the time while the index file is read. Now we read index
with progress, instead of doing it silently.

As a side-effect, on the first search progress goes 0..100 twice with
the same title, but it is better than the progress hang on 0.

The index reading is not cancelable.