better handling of unknown ELO in searchHeader
If user does not care about ELO of Black, then we should not care
whether Black ELO is known, that is allowUnknownElo is relevant only
if a range is specified for some parameter (say, Black ELO), and this
parameter is unknown.

Optimize searchHeader by checking only active ranges, for example, if
there is no ELO range specified then we should not check ELO at all.