prevent confusion with auto play
Notify the user that the autoplay is started and how to stop it (any interaction).

Notify the user that the autoplay is stopped.

Hint the user that onNextMoveLongClick does nothing in the end of variation.

At least on VirtualBox, onUserInteraction() is called on touch-up of
the long click that started autoplay, so we keep track of time when
the autoplay was started to ignore that spurious call.